+++ RELEASE_NOTES.md
@@ -21,6 +21,7 @@
 - Fixed when FramingAssistant was not opened before and a DSO was selected from the SkyAtlas as Framing Source an error could occur
 - Fixed scrolling through Framing Assistant Offline Sky Map while cursor was inside Rectangle ignored zooming
 - Fixed Alitude charts displaying wrong Twilight/Night predictions for some scenarios
+- Manual focus target list was not updating in some scenarios. Now it will always update. The interval for updates is one minute.
 
 ## Improvements
 - If telescope is capable of reporting SideOfPier there will now be a new option to consider this for calculating the need for meridian flips